hw4

+++ src/controllers/notesController.js
@@ -2,12 +2,14 @@ import createHttpError from 'http-errors';
 import  Note  from '../models/note.js';
 //get
 export const getAllNotes = async (req,res) =>{
-
+  const { _id: userId } = req.user;
   const {page=1,perPage=15,tag,search} = req.query;
   const skip = (page - 1) * perPage;
 
   const notesQuery = Note.find();
-
+  if (userId) {
+    notesQuery.where('userId').equals(userId);
+  }
   if(search){
     notesQuery.where({
   $or: [
@@ -33,7 +35,8 @@ export const getAllNotes = async (req,res) =>{
 
 export const getNoteById = async(req,res)=>{
   const {noteId} = req.params;
-  const note = await Note.findById(noteId);
+  const { _id: userId } = req.user;
+  const note = await Note.findOne({ _id: noteId, userId });
   if(!note){
     throw createHttpError(404, `Note not found`);
   }
@@ -41,13 +44,14 @@ export const getNoteById = async(req,res)=>{
 };
 //post
 export const createNote = async (req,res) =>{
-  const note = await Note.create(req.body);
+  const note = await Note.create({...req.body,userId: req.user._id,});
   res.status(201).json(note);
 };
 //delete
 export const deleteNote = async (req,res) =>{
   const {noteId} = req.params;
-  const note = await Note.findOneAndDelete({_id: noteId,});
+  const { _id: userId } = req.user;
+  const note = await Note.findOneAndDelete({_id: noteId,userId});
   if(!note){
     throw createHttpError(404, "Note not found");
   }
@@ -56,7 +60,8 @@ export const deleteNote = async (req,res) =>{
 //patch
 export const updateNote = async (req,res) =>{
   const {noteId} = req.params;
-  const note = await Note.findOneAndUpdate({ _id: noteId },req.body, { returnDocument: "after" },);
+  const { _id: userId } = req.user;
+  const note = await Note.findOneAndUpdate({ _id: noteId,userId },req.body, { returnDocument: "after" },);
   if(!note){
     throw createHttpError(404, "Note not found");
   }
